mock, @WebMvcTest

+++ intellij_workspace/springbootex13_board_mybatis/src/main/java/kr/co/jhta/app/springbootex13_board_mybatis/dto/BoardAttachedFileDTO.java
@@ -11,5 +11,5 @@ public class BoardAttachedFileDTO {
     private int fileNo;
     private String fileName;
     private String filePath;
-    private int boardNo;
+    private Long boardNo;
 }
